Migrate modes-resources concepts page to crag
