+++ docs/content-crag/_apidocs.mdx
@@ -16,7 +16,7 @@ APIs from the core `dagster` package, divided roughly by topic:
 
 - [Pipelines](/\_apidocs/pipeline) APIs to define pipelines, dependencies and fan-in dependencies between solids, and aliased instances of solids.
 
-- [Modes & Resources](/\_apidocs/modes-resources) APIs to define pipeline modes and resources.
+- [Resources](/\_apidocs/resources) APIs to define resources.
 
 - [Presets](/\_apidocs/presets) APIs to define configuration presets.
