+++ docs/content-crag/concepts.mdx
@@ -19,7 +19,7 @@ Each page in this section contains:
     to define orchestration graphs. This section covers how to define and use
     both solids and pipelines.
   </LinkGridItem>
-  <LinkGridItem title="Modes and Resources" href="/concepts/modes-resources">
+  <LinkGridItem title="Resources" href="/concepts/resources">
     Modes, alongside resources, enable you to separate the pipeline logic from
     its heavyweight external dependencies. This makes testing and develop data
     pipelines possible in various environments.
