+++ docs/content-crag/concepts/logging/loggers.mdx
@@ -9,14 +9,13 @@ Dagster includes a rich and extensible logging system. Dagster comes with a buil
 
 ## Relevant APIs
 
-| Name                                                  | Description                                                                                                                                            |
-| ----------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| <PyObject object="logger" decorator />                | The decorator used to define loggers. The decorator returns a <PyObject object="LoggerDefinition" />                                                   |
-| <PyObject object="LoggerDefinition" />                | Class for loggers. You almost never want to use initialize this class directly. Instead, you should use the decorator above                            |
-| <PyObject object="SolidExecutionContext" />           | The `context` object available to a solid compute function                                                                                             |
-| <PyObject module="dagster" object="ModeDefinition" /> | Modes allow you to vary pipeline behavior between different deployment environments. For more info, see the [Modes](/concepts/modes-resources) section |
-| <PyObject object="InitLoggerContext" />               | The context object passed to a custom logger's initialization function.                                                                                |
-| <PyObject object="build_init_logger_context"/>        | A function to construct a <PyObject object="InitLoggerContext" /> outside of execution, primarily to be used for testing purposes.                     |
+| Name                                           | Description                                                                                                                        |
+| ---------------------------------------------- | ---------------------------------------------------------------------------------------------------------------------------------- |
+| <PyObject object="logger" decorator />         | The decorator used to define loggers. The decorator returns a <PyObject object="LoggerDefinition" />                               |
+| <PyObject object="LoggerDefinition" />         | Class for loggers. You almost never want to use initialize this class directly. Instead, you should use the decorator above        |
+| <PyObject object="SolidExecutionContext" />    | The `context` object available to a solid compute function                                                                         |
+| <PyObject object="InitLoggerContext" />        | The context object passed to a custom logger's initialization function.                                                            |
+| <PyObject object="build_init_logger_context"/> | A function to construct a <PyObject object="InitLoggerContext" /> outside of execution, primarily to be used for testing purposes. |
 
 ## Overview
 
@@ -223,7 +222,7 @@ def test_init_json_console_logger_with_context():
 
 Logging is environment-specific: you don't want messages generated by data scientists' local development loops to be aggregated with production messages; on the other hand, you may find that in production console logging is irrelevant or even counterproductive.
 
-Dagster recognizes this by attaching loggers to [modes](/concepts/modes-resources) so that you can seamlessly switch from, e.g., Cloudwatch logging in production to console logging in development and test, without changing any of your code.
+Dagster recognizes this by attaching loggers to [modes](/concepts/resources) so that you can seamlessly switch from, e.g., Cloudwatch logging in production to console logging in development and test, without changing any of your code.
 
 ```python file=/concepts/logging/logging_modes.py startafter=start_logging_mode_marker_0 endbefore=end_logging_mode_marker_0
 @solid
